fixed event listeners
